+++ scripts/macros/monsterFeatures/dreamEater/dreamEater.js
@@ -1,0 +1,4 @@
+import {engulfingNightmare} from './engulfingNightmare.js';
+export let dreamEater = {
+    'engulfingNightmare': engulfingNightmare
+}