+++ scripts/macros/monsterFeatures/gibberingMouther/gibberingMouther.js
@@ -1,0 +1,4 @@
+import {gibbering} from './gibbering.js';
+export let gibberingMouther = {
+    'gibbering': gibbering
+}